feat: chore: refactor: new type system (refactoring structs.rs)
