+++ src/lib.rs
@@ -4,3 +4,4 @@ pub mod server;
 #[cfg(test)]
 mod tests;
 
+pub mod structs;
