+++ src/structs.rs
@@ -1,0 +1,6 @@
+pub mod config;
+pub mod context;
+pub mod keyboard;
+pub mod middleware;
+pub mod tg;
+pub mod vk;
